Move function into ShapeUtil.
